requirements.txt
numpy
opencv-python
rawpy
astropy
astroquery
imageio
Pillow